split LoadClusterConfiguration phase
