+++ lib/pharos/cluster_manager.rb
@@ -93,6 +93,7 @@ module Pharos
       apply_phase(Phases::ConfigureClient, [master_hosts.first], ssh: true, master: master_hosts.first, parallel: false)
 
       # master is now configured and can be used
+      apply_phase(Phases::LoadClusterConfiguration, [master_hosts.first], master: master_hosts.first)
       apply_phase(Phases::ConfigureDNS, [master_hosts.first], master: master_hosts.first)
 
       apply_phase(Phases::ConfigureWeave, [master_hosts.first], master: master_hosts.first) if config.network.provider == 'weave'
