+++ lib/pharos/phases/configure_client.rb
@@ -10,11 +10,6 @@ module Pharos
       def call
         fetch_kubeconfig
         client_prefetch
-        config_map = previous_config_map
-        return unless config_map
-
-        cluster_context['previous-config-map'] = config_map
-        cluster_context['previous-config'] = build_config(config_map)
       end
 
       def fetch_kubeconfig
@@ -28,22 +23,6 @@ module Pharos
       # prefetch client resources to warm up caches
       def client_prefetch
         kube_client.apis(prefetch_resources: true)
-      end
-
-      # @param configmap [K8s::Resource]
-      # @return [Pharos::Config]
-      def build_config(configmap)
-        cluster_config = Pharos::YamlFile.new(StringIO.new(configmap.data['cluster.yml']), override_filename: "#{@host}:cluster.yml").load
-        cluster_config['hosts'] ||= []
-        data = Pharos::ConfigSchema.build.call(cluster_config)
-        Pharos::Config.new(data)
-      end
-
-      # @return [K8s::Resource, nil]
-      def previous_config_map
-        kube_client.api('v1').resource('configmaps', namespace: 'kube-system').get('pharos-config')
-      rescue K8s::Error::NotFound
-        nil
       end
     end
   end
